+++ cli/azd/extensions/microsoft.azd.demo/internal/cmd/root.go
@@ -19,7 +19,6 @@ func NewRootCommand() *cobra.Command {
 	}
 
 	rootCmd.SetHelpCommand(&cobra.Command{Hidden: true})
-	rootCmd.PersistentFlags().Bool("debug", false, "Enable debug mode")
 
 	rootCmd.AddCommand(newListenCommand())
 	rootCmd.AddCommand(newContextCommand())